requirements.txt
langchain-openai~=0.3.17
langchain-core~=0.3.60
python-dotenv~=1.1.0
langchain~=0.3.25
pytest~=8.3.5